+++ build-binary-from-source.sh
@@ -1,4 +1,4 @@
 cd ..
-R CMD INSTALL --build --no-multiarch ooacquire_0.4.2.tar.gz
+R CMD INSTALL --build --no-multiarch ooacquire_0.4.2.9000.tar.gz
 cd ./ooacquire